+++ src/main/java/org/phoenix/planet/configuration/security/TokenAuthenticationFilter.java
@@ -14,8 +14,8 @@ import org.phoenix.planet.constant.AuthenticationError;
 import org.phoenix.planet.constant.TokenKey;
 import org.phoenix.planet.error.auth.TokenException;
 import org.phoenix.planet.provider.TokenProvider;
-import org.phoenix.planet.service.RefreshTokenService;
-import org.phoenix.planet.util.CookieUtil;
+import org.phoenix.planet.service.auth.RefreshTokenService;
+import org.phoenix.planet.util.cookie.CookieUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
