+++ src/main/java/org/phoenix/planet/controller/AuthController.java
@@ -10,8 +10,8 @@ import org.phoenix.planet.constant.AuthenticationError;
 import org.phoenix.planet.constant.TokenKey;
 import org.phoenix.planet.dto.member.request.SignUpRequest;
 import org.phoenix.planet.provider.TokenProvider;
-import org.phoenix.planet.service.MemberService;
-import org.phoenix.planet.util.CookieUtil;
+import org.phoenix.planet.service.member.MemberService;
+import org.phoenix.planet.util.cookie.CookieUtil;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
