+++ src/main/java/org/phoenix/planet/provider/TokenProvider.java
@@ -20,8 +20,8 @@ import org.phoenix.planet.constant.Role;
 import org.phoenix.planet.dto.auth.PrincipalDetails;
 import org.phoenix.planet.dto.member.raw.Member;
 import org.phoenix.planet.error.auth.TokenException;
-import org.phoenix.planet.service.RefreshTokenService;
-import org.phoenix.planet.util.CookieUtil;
+import org.phoenix.planet.service.auth.RefreshTokenService;
+import org.phoenix.planet.util.cookie.CookieUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
