+++ src/main/java/org/phoenix/planet/service/auth/RefreshTokenService.java
@@ -1,4 +1,4 @@
-package org.phoenix.planet.service;
+package org.phoenix.planet.service.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.phoenix.planet.repository.RefreshTokenRepository;
