+++ src/main/java/org/phoenix/planet/service/member/MemberService.java
@@ -1,4 +1,4 @@
-package org.phoenix.planet.service;
+package org.phoenix.planet.service.member;
 
 import org.phoenix.planet.dto.member.request.SignUpRequest;
 import org.springframework.web.multipart.MultipartFile;
